+++ test/wal-windows-test.el
@@ -29,122 +29,122 @@
       (was-called-with get-buffer-window (list 'buffer))
       (was-called select-window))))
 
-(defvar test-tab '(current-tab (name . "test-tab") (explicit-name . t) (wtb . "test-hash")))
+(defvar test-tab '(current-tab (name . "test-tab") (explicit-name . t) (wpr . "test-hash")))
 
-(ert-deftest wtb--create-hash-key ()
+(ert-deftest wpr--create-hash-key ()
   (with-mock ((random . (lambda () 42))
               (emacs-pid . (lambda () 1))
               (recent-keys . (lambda () 'keys))
               md5)
 
-    (wal-tab-buffers--create-hash-key "test")
+    (wal-partial-recall--create-hash-key "test")
     (was-called-with md5 (list "test421keys"))))
 
-(ert-deftest wtb--on-create--sets-cdr ()
-  (with-mock ((wal-tab-buffers--create-hash-key . (lambda (_) "test")))
+(ert-deftest wpr--on-create--sets-cdr ()
+  (with-mock ((wal-partial-recall--create-hash-key . (lambda (_) "test")))
 
     (let ((tab '(current-tab (name . "test-tab") (explicit-name))))
 
-      (wal-tab-buffers--on-create tab)
+      (wal-partial-recall--on-create tab)
 
-      (should (string= "test" (alist-get 'wtb tab))))))
+      (should (string= "test" (alist-get 'wpr tab))))))
 
-(ert-deftest wtb--key--returns-if-set ()
+(ert-deftest wpr--key--returns-if-set ()
   (with-mock ((tab-bar--current-tab . (lambda (&rest _) test-tab)))
-    (should (string= "test-hash" (wal-tab-buffers--key)))))
+    (should (string= "test-hash" (wal-partial-recall--key)))))
 
 (defmacro with-tab-history (&rest body)
   "Run BODY with a clear tab history and a temp buffer."
   (declare (indent 0))
   `(with-mock ((tab-bar--current-tab . (lambda (&rest _) test-tab)))
-     (let ((wal-tab-buffers--table (make-hash-table)))
+     (let ((wal-partial-recall--table (make-hash-table)))
        (with-temp-buffer
          ,@body))))
 
-(ert-deftest wtb--remember--remembers ()
+(ert-deftest wpr--remember--remembers ()
   (with-tab-history
 
-   (wal-tab-buffers--remember)
+   (wal-partial-recall--remember)
 
-   (should-not (null (gethash (alist-get 'wtb test-tab) wal-tab-buffers--table)))))
+   (should-not (null (gethash (alist-get 'wpr test-tab) wal-partial-recall--table)))))
 
-(ert-deftest wtb--remember--inserts-once ()
-  (defvar wal-tab-buffers--table)
+(ert-deftest wpr--remember--inserts-once ()
+  (defvar wal-partial-recall--table)
 
   (with-tab-history
 
-    (wal-tab-buffers--remember)
-    (wal-tab-buffers--remember)
+    (wal-partial-recall--remember)
+    (wal-partial-recall--remember)
 
-    (should (eq 1 (ring-length (gethash (alist-get 'wtb test-tab) wal-tab-buffers--table))))))
+    (should (eq 1 (ring-length (gethash (alist-get 'wpr test-tab) wal-partial-recall--table))))))
 
-(ert-deftest wtb--forget--forgets ()
+(ert-deftest wpr--forget--forgets ()
     (with-tab-history
-      (wal-tab-buffers--remember)
+      (wal-partial-recall--remember)
 
-      (wal-tab-buffers--forget)
+      (wal-partial-recall--forget)
 
-      (should (eq 0 (ring-length (gethash (alist-get 'wtb test-tab) wal-tab-buffers--table))))))
+      (should (eq 0 (ring-length (gethash (alist-get 'wpr test-tab) wal-partial-recall--table))))))
 
-(ert-deftest wtb--on-close ()
+(ert-deftest wpr--on-close ()
   (with-tab-history
-    (wal-tab-buffers--remember)
+    (wal-partial-recall--remember)
 
-    (should (eq 1 (length (hash-table-keys wal-tab-buffers--table))))
+    (should (eq 1 (length (hash-table-keys wal-partial-recall--table))))
 
-    (wal-tab-buffers--on-close test-tab t)
+    (wal-partial-recall--on-close test-tab t)
 
-    (should (eq 1 (length (hash-table-keys wal-tab-buffers--table))))
+    (should (eq 1 (length (hash-table-keys wal-partial-recall--table))))
 
-    (wal-tab-buffers--on-close test-tab nil)
+    (wal-partial-recall--on-close test-tab nil)
 
-    (should (eq 0 (length (hash-table-keys wal-tab-buffers--table))))))
+    (should (eq 0 (length (hash-table-keys wal-partial-recall--table))))))
 
-(ert-deftest wtb--history ()
+(ert-deftest wpr--history ()
   (with-tab-history
-    (should-not (wal-tab-buffers--history))
+    (should-not (wal-partial-recall--history))
 
-    (wal-tab-buffers--remember)
+    (wal-partial-recall--remember)
 
-    (should (wal-tab-buffers--history))))
+    (should (wal-partial-recall--history))))
 
-(ert-deftest wtb--current-p ()
+(ert-deftest wpr--current-p ()
   (with-tab-history
-    (wal-tab-buffers--remember)
-    (should (wal-tab-buffers--current-p (current-buffer)))))
+    (wal-partial-recall--remember)
+    (should (wal-partial-recall--current-p (current-buffer)))))
 
-(ert-deftest wtb--has-buffers-p ()
+(ert-deftest wpr--has-buffers-p ()
   (with-tab-history
-    (should-not (wal-tab-buffers--has-buffers-p))
+    (should-not (wal-partial-recall--has-buffers-p))
 
-    (wal-tab-buffers--remember)
+    (wal-partial-recall--remember)
 
-    (should (wal-tab-buffers--has-buffers-p))))
+    (should (wal-partial-recall--has-buffers-p))))
 
-(ert-deftest wtb--known-buffer-p ()
+(ert-deftest wpr--known-buffer-p ()
   (with-tab-history
-    (wal-tab-buffers--remember)
+    (wal-partial-recall--remember)
 
-    (should (wal-tab-buffers--known-buffer-p (current-buffer)))))
+    (should (wal-partial-recall--known-buffer-p (current-buffer)))))
 
-(ert-deftest wtb--maybe-remember ()
+(ert-deftest wpr--maybe-remember ()
   (with-tab-history
-    (with-mock (wal-tab-buffers--remember
+    (with-mock (wal-partial-recall--remember
                 (buffer-file-name . #'always)
                 (buffer-list . (lambda () '(buffer)))
-                (wal-tab-buffers--known-buffer-p . #'ignore))
+                (wal-partial-recall--known-buffer-p . #'ignore))
 
-      (wal-tab-buffers--maybe-remember)
-      (was-called wal-tab-buffers--remember))))
+      (wal-partial-recall--maybe-remember)
+      (was-called wal-partial-recall--remember))))
 
-(ert-deftest wtb--on-frame-delete ()
+(ert-deftest wpr--on-frame-delete ()
   (defvar tab-bar-tabs-function nil)
 
   (let ((tab-bar-tabs-function (lambda (_) '(one two))))
-    (with-mock (wal-tab-buffers--on-close)
+    (with-mock (wal-partial-recall--on-close)
 
-      (wal-tab-buffers--on-frame-delete 'frame)
+      (wal-partial-recall--on-frame-delete 'frame)
 
-      (was-called-n-times wal-tab-buffers--on-close 2))))
+      (was-called-n-times wal-partial-recall--on-close 2))))
 
 ;;; wal-windows-test.el ends here